python-agent/requirements.txt
flask==2.3.3
flask-cors==4.0.0
groq==0.19.0
phidata==2.7.10
dotenv==0.9.9
python-dotenv==1.0.0
duckdb==1.2.1
setuptools==76.0.0
youtube-transcript-api==1.0.1
wikipedia==1.4.0